Configure with docker-compose
Added `example.env` as an example of what `.env` should look like. Devs
can fill in the empty values as appropriate.

Added `docker-compose.yml` with a mongo service and the Activist app
running as the web service. This makes it easy to build and spin up the
docker containers with appropriate links.

Updated README.